refactor NestedWizard class and update arch doc
